+++ bpm-engine-api/src/main/java/io/takari/bpm/api/Engine.java
@@ -9,13 +9,22 @@ public interface Engine {
 
     /**
      * Starts a new process instance with the given ID.
-     * @param processBusinessKey an external process instance ID, must be
-     * unique.
+     * @param processBusinessKey an external process instance ID, must be unique per {@link Engine}.
      * @param processDefinitionId the id of the process definition, can't be null.
-     * @param variables variables to be passed, can be null.
+     * @param arguments arguments to be passed, can be null.
      * @throws ExecutionException
      */
-    void start(String processBusinessKey, String processDefinitionId, Map<String, Object> variables) throws ExecutionException;
+    void start(String processBusinessKey, String processDefinitionId, Map<String, Object> arguments) throws ExecutionException;
+
+    /**
+     * Starts a new process instance with the given ID.
+     * @param processBusinessKey an external process instance ID, must be unique per {@link Engine}.
+     * @param processDefinitionId the id of the process definition, can't be null.
+     * @param variables initial variables
+     * @param arguments extra arguments, can be null.
+     * @throws ExecutionException
+     */
+    void start(String processBusinessKey, String processDefinitionId, Variables variables, Map<String, Object> arguments) throws ExecutionException;
 
     /**
      * Resumes an process instance, waiting for the specific event.